+++ src/backend/marsha/core/tests/utils/test_send_emails.py
@@ -6,6 +6,7 @@ from django.utils.datetime_safe import datetime
 
 from marsha.core.factories import VideoFactory
 from marsha.core.utils.send_emails import (
+    send_convert_reminder_notification,
     send_ready_to_convert_notification,
     send_vod_ready_notification,
 )
@@ -137,6 +138,49 @@ Access the VOD from Marsha [https://marsha.education/my-contents/videos/{video.i
 Or
 
 Access the VOD through the live stream from your LMS [{video.last_lti_url}]
+
+
+
+
+
+This mail has been sent to {email} by Marsha""".splitlines(),
+        )
+
+    def test_send_convert_reminder_notification(self):
+        """Test send_convert_reminder_notification."""
+        video = VideoFactory(
+            starting_at=datetime(2020, 1, 1, 9, 30, tzinfo=timezone.utc),
+            live_info={
+                "live_stopped_with_email": "user@example.com",
+            },
+            last_lti_url="https://example.com/course/machin",
+        )
+
+        send_convert_reminder_notification(video)
+
+        email = video.live_info.get("live_stopped_with_email")
+        self.assertEqual(mail.outbox[0].to[0], email)
+        self.assertEqual(
+            mail.outbox[0].subject,
+            f"Live Stream will be deleted soon - {video.title}",
+        )
+        self.assertListEqual(
+            mail.outbox[0].body.splitlines(),
+            f"""
+Live Stream will be deleted soon
+
+{video.title}
+
+Hello,
+The live stream you recorded for "{video.title}" is ready to be converted to VOD.
+The live stream recording will be available until Jan. 15, 2020, 9:30 a.m..
+
+Access the live stream from Marsha [https://marsha.education/my-contents/videos/{video.id}]
+
+
+Or
+
+Access the live stream from your LMS [{video.last_lti_url}]