+++ src/backend/marsha/development/management/commands/test_send_emails.py
@@ -39,3 +39,25 @@ class Command(BaseCommand):
             self.stdout.write("Convert to vod ready notification email sent")
         else:
             self.stderr.write("Convert to vod ready notification email not sent")
+
+        self.stdout.write()
+        self.stdout.write("Convert to vod reminder")
+        if send_convert_reminder_notification(video):
+            self.stdout.write("Convert to vod reminder notification email sent")
+        else:
+            self.stderr.write("Convert to vod reminder notification email not sent")
+
+        self.stdout.write()
+        self.stdout.write("Ready to convert notification (no LTI URL)")
+        video = VideoFactory(
+            starting_at=timezone.now() - timezone.timedelta(days=1),
+            live_info={
+                "live_stopped_with_email": "user@example.com",
+            },
+        )
+        if send_ready_to_convert_notification(video):
+            self.stdout.write("Ready to convert notification (no LTI URL) email sent")
+        else:
+            self.stderr.write(
+                "Ready to convert notification (no LTI URL) email not sent"
+            )