+++ src/backend/marsha/settings.py
@@ -469,6 +469,7 @@ class Base(Configuration):
     # LIVE SETTINGS
     NB_DAYS_BEFORE_DELETING_LIVE_RECORDINGS = values.Value(14)
     NB_DAYS_BEFORE_DELETING_AWS_ELEMENTAL_STACK = values.Value(14)
+    NB_DAYS_BEFORE_DELETING_AWS_ELEMENTAL_STACK_REMINDER = values.Value(3)
     NB_SECONDS_LIVING_DEV_STACK = values.PositiveIntegerValue(600)
     LIVE_PLAYLIST_WINDOW_SECONDS = values.PositiveIntegerValue(10)
     LIVE_SEGMENT_DURATION_SECONDS = values.PositiveIntegerValue(4)