+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Send an email when a live stream is ready to be converted to VOD
 - Send an email when a live stream is converted to VOD
+- Send an email when a live stream is about to be deleted
 
 ### Changed